+++ components/grid/index.tsx
@@ -8,7 +8,7 @@ const thumbnailHeight = 768;
 type Props = {
     isThree?: boolean;
     isOne?: boolean;
-    data: any;
+    data: IData;
 };
 
 const Grid: React.FC<Props> = ({ data, isThree, isOne }) => {
@@ -229,5 +229,45 @@ const Grid: React.FC<Props> = ({ data, isThree, isOne }) => {
         </div>
     );
 };
+
+interface IData {
+    sort(arg0: (a: any, b: any) => 0 | 1 | -1);
+    acf: {
+        data: string;
+        imatge_destacada: {
+            widht: string;
+            height: string;
+        };
+    };
+    type: string;
+    kind: string;
+    media: string;
+    snippet: {
+        publishedAt: string;
+        title: string;
+        channelTitle: string;
+        thumbnails: {
+            high: {
+                url: string;
+            };
+        };
+    };
+    dateupload: string;
+    title:
+        | string
+        | {
+              rendered: string;
+          };
+    id:
+        | string
+        | {
+              videoId: string;
+          };
+    slug: string
+    datetaken: string
+    url_m: string
+    width_m: string
+    height_m: string
+}
 
 export default Grid;
