+++ components/mdncontentparser/index.tsx
@@ -1,10 +1,16 @@
 import Markdown from 'markdown-to-jsx';
 import styles from './MDFileContent.module.scss';
 
-export default function MDFileContent({ content }) {
+type Props = {
+    content: string;
+};
+
+const MDFileContent: React.FC<Props> = ({ content }) => {
     return (
         <div className={styles.MDFileContentComponent}>
             <Markdown>{content}</Markdown>
         </div>
     );
-}
+};
+
+export default MDFileContent;
