+++ pages/8-de-juny-de-1969.tsx
@@ -2,26 +2,29 @@ import Layout from '@components/layout';
 import styles from '@styles/Home.module.scss';
 import Figure from '@components/figure';
 import api from '@libs/api.js';
-import MDFileContent from '@components/mdncontentparser';
 import OtherRoutes from '@components/otherroutes';
+import MDFileContent from '@components/mdncontentparser';
+import { GetStaticProps } from 'next';
 
 
 const staticDataUrl = process.env.STATIC_DATA_URL;
 
-const CinquantaAnys50Tuits = ({ anys50tuits, footer, mdFileContent, routes }) => {
-    const { title, pageTitle, pageDescription, otherRoutes } = anys50tuits.meta;
+const VuitDeJunyDe1969 = ({ vuitDeJuny1969, footer, mdFileContent, routes }) => {
+    const { title, pageTitle, pageDescription, otherRoutes } = vuitDeJuny1969.meta;
     const { routes: footerLinks, supporters } = footer;
-    const imageGallery = anys50tuits.images.imageGallery;
+    const mainImage = vuitDeJuny1969.images.mainImage;
+    const imageGallery = vuitDeJuny1969.images.imageGallery;
     return (
         
             <Layout
                 pageTitle={pageTitle}
                 title={title}
-                navRoutes={routes}
                 pageDescription={pageDescription}
                 footerLinks={footerLinks}
                 supporters={supporters}
+                navRoutes={routes}
             >
+                <Figure data={mainImage} quality={100} layout={'responsive'} />
 
                 <div className={`${styles.container}`}>
                     <main className={styles.main}>
@@ -37,18 +40,17 @@ const CinquantaAnys50Tuits = ({ anys50tuits, footer, mdFileContent, routes }) =>
     );
 };
 
-export const getStaticProps = async () => {
-    const [anys50tuits, footer, routes] = await Promise.all([
-        api.anys50tuits.getData(),
+export const getStaticProps: GetStaticProps = async () => {
+    const [vuitDeJuny1969, footer, routes] = await Promise.all([
+        api.vuitDeJuny1969.getData(),
         api.footer.getData(),
         api.routes.getData(),
     ]);
-    const res = await fetch(`${staticDataUrl}/content/50-anys-50-tuits.md`);
+    const res = await fetch(`${staticDataUrl}/content/1969.md`);
     const mdFileContent = await res.text();
-
     return {
         props: {
-            anys50tuits: { ...anys50tuits[0] },
+            vuitDeJuny1969: { ...vuitDeJuny1969[0] },
             footer: { ...footer[0] },
             mdFileContent: mdFileContent,
             routes,
@@ -56,4 +58,4 @@ export const getStaticProps = async () => {
     };
 };
 
-export default CinquantaAnys50Tuits;
+export default VuitDeJunyDe1969;
