+++ pages/_app.tsx
@@ -1,6 +1,7 @@
 import '@styles/globals.scss';
+import type { AppProps } from 'next/app';
 
-function MyApp({ Component, pageProps }) {
+function MyApp({ Component, pageProps }: AppProps) {
     return <Component {...pageProps} />;
 }
