+++ pages/fes-castells.tsx
@@ -3,49 +3,31 @@ import styles from '@styles/Home.module.scss';
 import api from '@libs/api.js';
 import Figure from '@components/figure';
 import { FaMapSigns } from 'react-icons/fa';
-import { MailruIcon } from 'react-share';
+import { GetStaticProps } from 'next';
 
-const Contacte = ({ contacte, footer, routes }) => {
-    const {
-        title,
-        pageTitle,
-        pageDescription,
-        name,
-        address,
-        phone,
-        mobile,
-        web,
-        email,
-        map,
-    } = contacte.meta;
-    const mainImage = contacte.images.mainImage;
+const FesCastells = ({ footer, routes, fesCastells }) => {
+    const { title, pageTitle, pageDescription, map } = fesCastells.meta;
     const { routes: footerLinks, supporters } = footer;
+    const mainImage = fesCastells.images.mainImage;
     return (
         <Layout
             title={title}
             pageTitle={pageTitle}
             pageDescription={pageDescription}
-            contacte
             navRoutes={routes}
             footerLinks={footerLinks}
             supporters={supporters}
         >
             <div className={`${styles.container} ${styles.withOverlay}`}>
                 <main className={`${styles.main} ${styles.withUnderlay}`}>
-                    <p className={styles.description}>
-                        <strong>{name}</strong>
-                        <br />
-                        <br />
-                        {address}
-                        <br />
-                        T. <a href={phone.href}>{phone.number}</a>
-                        <br />
-                        M. <a href={mobile.href}>{mobile.number}</a> [premsa]
-                        <br />
-                        {web}
-                        <br />
-                        <a href={email.href}>{email.address}</a>
-                        <br />
+                    <h1>... Arribaran!</h1>
+                    <h2
+                        className={styles.title}
+                        dangerouslySetInnerHTML={{
+                            __html: pageDescription,
+                        }}
+                    />
+                    <h3 className={styles.title}>
                         <a
                             title={map.title}
                             target={'_blank'}
@@ -54,7 +36,8 @@ const Contacte = ({ contacte, footer, routes }) => {
                         >
                             <FaMapSigns />
                         </a>
-                    </p>
+                    </h3>
+                    <p>(quan la COVID-19 deixi d'emprenyar)</p>
                 </main>
             </div>
             <Figure data={mainImage} quality={100} layout={'responsive'} />
@@ -62,19 +45,19 @@ const Contacte = ({ contacte, footer, routes }) => {
     );
 };
 
-export const getStaticProps = async () => {
-    const [contacte, footer, routes] = await Promise.all([
-        api.contacte.getData(),
+export const getStaticProps: GetStaticProps = async () => {
+    const [fesCastells, footer, routes] = await Promise.all([
+        api.fesCastells.getData(),
         api.footer.getData(),
         api.routes.getData(),
     ]);
     return {
         props: {
-            contacte: { ...contacte[0] },
             footer: { ...footer[0] },
+            fesCastells: { ...fesCastells[0] },
             routes,
         },
     };
 };
 
-export default Contacte;
+export default FesCastells;
